FossilOrigin-Name: 1a5181b53d9f1ee1d8d57b4d57214b0d5b11e8879a48b0b0f34e9b2a439332ef

+++ src/sqliteInt.h
@@ -1808,7 +1808,6 @@ struct sqlite3 {
   int bHctMigrate;
 };
 
-
 /*
 ** Candidate values for sqlite3.eConcurrent
 */
@@ -1973,7 +1972,7 @@ struct sqlite3 {
 ** field is used by per-connection app-def functions.
 */
 struct FuncDef {
-  i8 nArg;             /* Number of arguments.  -1 means unlimited */
+  i16 nArg;            /* Number of arguments.  -1 means unlimited */
   u32 funcFlags;       /* Some combination of SQLITE_FUNC_* */
   void *pUserData;     /* User data parameter */
   FuncDef *pNext;      /* Next function with same name */
